+++ src/custom/dbFirebase.js
@@ -1,6 +1,9 @@
+import firebase from 'firebase/app';
+
 import firebaseApp from './firebaseInit.js';
 
-const db = firebaseApp.firestore();
+const firestore = firebaseApp.firestore();
+const storageRef = firebaseApp.storage().ref();
 
 async function fetchPhotos() {
   const geojson = {
@@ -8,7 +11,7 @@ async function fetchPhotos() {
     "features": []
   };
 
-  const querySnapshot = await db.collection("photos").get();
+  const querySnapshot = await firestore.collection("photos").get();
 
   querySnapshot.forEach( doc => {
       console.log(`${doc.id} =>`, doc.data());
@@ -35,9 +38,30 @@ async function fetchPhotos() {
   return geojson;
 }
 
+async function saveMetadata(data) {
+  if (firebase.auth().currentUser) {
+    data.owner_id = firebase.auth().currentUser.uid;
+  }
+  data.updated = firebase.firestore.FieldValue.serverTimestamp();
+  const photoRef = await firestore.collection('photos').add(data);
+
+  return photoRef
+}
+
+async function savePhoto(id, base64) {
+  const originalJpgRef = storageRef.child("photos").child(id).child("original.jpg");
+  console.log(base64)
+  return await originalJpgRef.putString(base64, "base64", {contentType:"image/jpeg"});
+}
+
+async function uploadPhoto(data) {
+  const photoRef = await saveMetadata({location: new firebase.firestore.GeoPoint(data.latitude, data.longitude), description: data.text  });
+  return await savePhoto(photoRef.id, data.base64);
+}
+
 async function getUser(id) {
-  const fbUser = await db.collection("users").doc(id).get();
+  const fbUser = await firestore.collection("users").doc(id).get();
   return fbUser.exists ? fbUser.data() : null;
 }
 
-export default {fetchPhotos, getUser};
+export default {fetchPhotos, getUser, uploadPhoto};