+++ src/custom/firebaseInit.js
@@ -5,6 +5,8 @@ import 'firebase/auth';
 import 'firebase/firestore';
 // import 'firebase/messaging';
 // import 'firebase/functions';
+import 'firebase/storage';
+
 
 // Initialize Firebase
 const config = {