+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - introducing configurable components
 - get user groups
 - moderator user
+- upload images
 
 ## 0.2.2 (2018-11-05)
 - fix map iOS